requirements.txt.txt
scikit-learn==1.2.2
pickle
flasks